Add change detection logging to the mtime and inotify file watchers
Summary:
The mtime and inotify file watchers will now emit a log message containing the
path of the changed file that is triggering a reload.

I'll do the same for the fsevents file watcher when I restore it but I'm not
going to touch the win32 file watcher.

Test Plan:
I tested the mtime file watcher on my Mac.

1. Started dev_appserver.py with the usual arguments for serving webapp
2. Ran `touch Makefile` and saw the log message "Detected change in Makefile"
3. Deleted and re-added the Makefile and saw the same log message for both
4. Ran `touch content/models.py` and saw log messages for content, content/models.py, and genfiles/kake-server.log

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D9084