Restore the FSEvents-based file watcher, and remove the WinFileWatcher.
This is a cherry-pick from the 1.9.6 branch:
   3cf5710e777550cbf1b684d580da0a5f65f27b6c
   3fd0b77bed73aa7036c70bf0c4e3cfff4fd170f0
with the following modification for 1.9.10:

Instead of adding a new watcher_common.ignore_dir() method, I use the
existing (in 1.9.10) skip_ignored_dirs() method.

We need to check *every* directory component, not just the lowest one.

skip_ignored_dirs() is weird in that it takes just a single directory
name, not a full path (so it will be useless for our skip-files regexp
stuff one day).  Ah well.

As part of this, we needed to update the `changes` code since it looks
like the `changes` spec was changed to have a long polling behavior
with the GAE upgrade: it takes a `timeout_ms` and is expected to block
until either any files have changed or the timeout expires.  (Instead,
we were always returning the set of changes immediately, even if it
was empty, so the dev_appserver code was calling it continuously in a
loop.)  This commit changes the implementation so `_fsevents_callback`
notifies an event so `changes` returns when the changes actually
happen (or after the timeout).  As for the timeout arg, I //think//
it's correct to ignore it, since fsevents file-watching is continuous
and in its own thread.  But I don't know for sure.  *shrug*.

Auditors: dylan, benkraft

Test Plan:
Dylan will test!

Stop dev_appserver from using 100% CPU with fsevents_file_watcher

Summary:
https://app.asana.com/0/27216215224639/52930358597729

It looks like the `changes` spec was changed to have a long polling behavior
with the GAE upgrade: it takes a `timeout_ms` and is expected to block until
either any files have changed or the timeout expires. Instead, we were always
returning the set of changes immediately, even if it was empty, so the
dev_appserver code was calling it continuously in a loop. This commit changes
the implementation so `_fsevents_callback` notifies an event so `changes`
returns when the changes actually happen (or after the timeout).

Test Plan:
Run dev_appserver with webapp. Change main.py to add a print statement and
save. The logs show that the file gets reloaded. Look in Activity Monitor. There
is no longer a Python process using 100% CPU.

Reviewers: benkraft, csilvers

Reviewed By: csilvers

Subscribers: riley

Differential Revision: https://phabricator.khanacademy.org/D22181