Have the file-watchers ignore files that are not in skip-files.
Summary:
The motivation is that if a file is in an app.yaml's skip_files, then
it is not accessible to the dev-appserver, and thus we don't care if
it changes or not.

This allows applications to store files such as the datastore inside
the app directory, without having to worry about constant reloads as a
result.  It also opens the door -- if we wanted to -- of lifting the
assumption that we never need to watch dot-directories in
watcher_common.py, and instead explicitly skip the dot-directories
that the user puts in skip_files.

As part of this change, I removed the _IMMUTABLE_PROPERTIES part of
ModuleConfiguration.  It didn't work well with skip-files reloading,
and I believe that's obsolete these days in any case, since we always
spin up a new instance when the yaml files change.

The implementation mostly involves passing around the skip-files RE to
the watchers.  We have to do a little work on some watchers to get the
filenames relative to application-root, since the filenames in the
skip-files RE are all taken to be relative to the application-root.

Test Plan:
   env PYTHONPATH=. python google/appengine/tools/devappserver2/inotify_file_watcher_test.py
   env PYTHONPATH=. python google/appengine/tools/devappserver2/mtime_file_watcher_test.py
   env PYTHONPATH=. python google/appengine/tools/devappserver2/win32_file_watcher_test.py

Reviewers: chris, dylan

Reviewed By: chris, dylan

Subscribers: dylan

Differential Revision: http://phabricator.khanacademy.org/D13232