Add patch for python 2.7.9 compatibility when using the remote API.
See the patch at: https://code.google.com/p/googleappengine/issues/detail?id=11536.

Note that this is not intended to be merged into master or substated into
webapp.  We should fix this correctly by updating the SDK in our fork, but this
allows python 2.7.9 users to use the remote API when needed by checking out a
different branch.

Test Plan:
- from webapp, `tools/devshell.py --prod`, verify that devshell starts.