Additional patch for python 2.7.9 compatibility.
This is based on the patch in d7d4bdb6, where extra kwargs were required in
urllib2 functions because the API changed in 2.7.9.  This commit additionally
changes the httplib functions in the SDK's override of httplib to be compatible
with 2.7.9.

Note that this is not intended to be merged with master or substated into
webapp.