Stop dev_appserver from using 100% CPU with fsevents_file_watcher
Summary:
https://app.asana.com/0/27216215224639/52930358597729

It looks like the `changes` spec was changed to have a long polling behavior
with the GAE upgrade: it takes a `timeout_ms` and is expected to block until
either any files have changed or the timeout expires. Instead, we were always
returning the set of changes immediately, even if it was empty, so the
dev_appserver code was calling it continuously in a loop. This commit changes
the implementation so `_fsevents_callback` notifies an event so `changes`
returns when the changes actually happen (or after the timeout).

Test Plan:
Run dev_appserver with webapp. Change main.py to add a print statement and
save. The logs show that the file gets reloaded. Look in Activity Monitor. There
is no longer a Python process using 100% CPU.

Reviewers: benkraft, csilvers

Reviewed By: csilvers

Subscribers: riley

Differential Revision: https://phabricator.khanacademy.org/D22181