Two fixes to get things running better under PyPy.
Summary:
GAE wants to do this so it can replace things like `thread` with its own
versions. On anything but CPython, that'll be double plus ungood, so
don't do that.

Also got rid of some dodgy code so it works with pypy2.4 (and does not
warn on pypy4.0).

Test Plan: ran locally

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D25533