Pass down dev appserver restart time to python runtime
Summary:
We'd like to have caches that depend on the last `dev_appserver` restart time,
as a simulation of depending on the App Engine version in prod.  (For example,
we want to have caches that are expired on a new App Engine version expire
whenever the dev server restarts.)  `dev_appserver`'s own version is set on
first start, so that won't change when we want it to, and `dev_appserver` may
start multiple instances, so doing something in our code at import-time will
change too often.  So we need to add our own.

Figuring out the right place to do this was a little tricky, but it seemed like
there was no way to do it generally for all runtimes, since we need to hook
into instance creation to edit the runtime config and add the environment
variable.  So we do it in the python runtime, and just have it log the last
time it restarted (i.e. either startup or on file-change).  Then we can just
inject it into the runtime config's environment.

Test Plan:
Added an import-time print statement in our code (`ka_globals.py`) to print
`os.environ['DEV_APPSERVER_LAST_RESTART_TIME']`.  Started a dev server (with
`--max_module_instances=10` to make sure it would load a bunch of new
instances), and saw it print the same value several times, even after loading
some pages.  Touched a file, and saw it print the same new value several times.
Restarted the dev server with `--automatic_restarts 0`, and saw it print the
same value several times, and not change even after touching some files and
loading some pages.

Reviewers: johnsullivan, csilvers

Reviewed By: csilvers

Subscribers: #chameleon

Differential Revision: https://phabricator.khanacademy.org/D32797