Separate out key-translation tests, and add a bunch more
Summary:
Writing tests has often been a good exercise so far, and it continues to
be: this found another minor bug (passing an invalid int would give a
server error).  This will also be useful as there will be a lot more
keys to translate in future diffs.  I left in a smattering of tests for
allocate IDs to test the handler wrappers.

Finally, I added a shortcut script to run all of the tests we've added
to frankenserver -- sadly running all of Google's tests doesn't work but
we can at least run all of ours.

Test Plan: `python/run_tests.py --frankenserver`

Reviewers: csilvers

Subscribers: aasmund, colin

Differential Revision: https://phabricator.khanacademy.org/D53749