Implement lookup (datastore Get)
Summary:
In this commit I implement the 'lookup' API (which is what datastore
calls 'Get').  The bulk of the commit is the translation of App Engine
entities (`datastore.Entity`) into REST format, which includes
translating each of the possible property types.

Many of these (ints, strings, etc.) were fairly obvious from the
documentation.  Others (like `UserProperty`) I just had to look in the
real datastore to see what the prod API returned.  Then, it was just a
matter of gluing it all together.

Of interest are two things we handle differently from REST: "meaning"
and "version".  The former is a legacy notion, which we include
potentially more than necessary because we're not sure when it's correct
to do so; the second we set to a bogus value because it's not exposed
via the App Engine API.  Neither seems to be used by clients.

The tests were a bunch of work.  Again, I tried to do the bulk of the
testing at the entity-translation level rather than the API call level.
But there are a lot of properties to test, and I did my best to verify
that each matches what prod does.  In many cases, I could just find some
model using the given property, hit the prod API, and see what it gave
me.  (I made especially sure to check all the string/bytes types, since
those can get messy.)

Test Plan: `python/run_tests.py --frankenserver`

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: aasmund, colin

Differential Revision: https://phabricator.khanacademy.org/D53808