Separate out value-translation from entity-translation
Summary:
The bulk of entity translation is actually in translating (property)
values, and that will be shared with query resolution.  So I broke it
out into a separate file.  Sadly, `entityValue` makes for a circular
import, but otherwise it's a very clean break.

Test Plan: `tools/run_tests.py --frankenserver`

Reviewers: csilvers

Subscribers: aasmund, colin

Differential Revision: https://phabricator.khanacademy.org/D54000