Implement value translation from REST back to GAE (mostly)
Summary:
In this commit, I implement the other direction of translating
(property) values -- from REST back to GAE.  This will be used for puts
as well as for queries; both will come in future commits.  I did not yet
implement the toplevel entity translation; I'll do that along with puts
but I will likely implement queries first.

Most of this was just a matter of reversing all the translators I
already had (and refactoring a bit so that things still make sense).
It's a little weird because we have to do some guessing as to datastore
types, but mostly I think it doesn't matter if we get it perfect.  The
only truly messy case was users and entities: as far as I can tell
`UserProperty` is represented by REST in exactly the same way as a
`LocalStructuredProperty` with a particular set of properties, so we
just have to guess.  In practice it should be fine.

Finally, I added some quick tests for value translation.  The real test
will be when I implement put and query.  I also fixed a bug in the other
direction of translation which I noticed while implementing this one.

Test Plan: `python/run_tests.py --frankenserver`

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: colin, aasmund

Differential Revision: https://phabricator.khanacademy.org/D54051