Implement queries in the datastore translator
Summary:
In this commit, I implement the `runQuery` call, which corresponds to
GAE's `datastore.Query().Run()`.  This was in some ways pretty simple --
most of the groundwork was already in place -- and in some ways a bit of
a doozy.  It's the last remaining read call; after this it should be
possible to wire up services that only want to do reads (like
kotlin-routes) to the translator for some real live testing.

Queries are a fairly complicated API, so there's a lot of work to just
massage the fields that come in from REST into the form datastore wants.
To keep things organized I broke this out into a separate file, although
sadly it's hard to separate the translation from the execution like I
did for gets.  Most of this stuff is pretty straightforward, though.
Similarly, the tests are very verbose, even with some helpers, but there
shouldn't be any surprises there.

I did a few small refactors to support shared logic: I broke out the
read-option-parsing into a helper, as well as the namespace translation
and project ID validation.

Test Plan: `python/run_tests.py --frankenserver`

Reviewers: csilvers

Subscribers: yansong, aasmund, colin

Differential Revision: https://phabricator.khanacademy.org/D54249