Implement protobuf-over-HTTP form of Datastore API
When I originally planned this datastore translator, I made sure to
check that the API clients used, or at least could use, the HTTP form of
the API, rather than gRPC, so I could avoid having to set up the latter.
Thus, I implemented the REST API as documented, using JSON requests and
responses.  But it turns out my search did not go far enough: actually
what all the clients use is a protobuf-over-HTTP form of the API!

We now support both.  I had to figure out some of the details of the
protobuf-over-HTTP (sadly it's not documented, although mostly it's just
a direct translation) and importing the genprotos into frankenserver was
a bit of a mess (although luckily the proto library itself was already
available, with one minor fix).  In principle, it might be cleaner to
convert all the code to use the protos, but it's already written to
expect JSON, and in python it's not like we'd get type safety anyway.

Test plan:
I didn't try to write unit tests for this change -- I worry I'd just be
testing that the translation works as I think it does.  (Although
existing tests of course still pass.)  Instead, I wired the
`kotlin-routes` service in webapp up to the translator by running:
- in webapp root: `make serve FLAGS=--enable_datastore_translator`
- in webapp `services/kotlin-routes`:
  `DATASTORE_EMULATOR_HOST=localhost:8001 ./gradlew appengineRun`
and then going to `localhost:8081`, clicking around, and watching the
latter console for errors.  (In a separate commit, I'll set up that
wiring more automatically.)  I also modified `kotlin-routes` to print
the annotation proto before sending it (in `addServerSetFieldsAndBuild`
-- sadly doing it where we mock pub/sub is difficult) to see that we
were sending reasonable annotations; I made sure to check that I could
get it to send assignment annotations (which require datastore queries)
as well as the correct user.

Reviewers: csilvers

Subscribers: yansong, aasmund, colin