Fix up import problems with protobuf (used by the datastore translator)
As described in D54474, the datastore translator uses protobuf to
implement the protobuf-over-HTTP version of the datastore REST API.  For
me, it works just fine to use the existing vendored copy of protobuf in
`python/lib/grpcio-1.9.1`.  But sadly, this doesn't work for some folks:
devappserver only adds this directory to the path if it cam import
`grpc` successfully, and its copy of `grpc` isn't built for most
platforms.  (Or something.  Neither mac users nor some other linux users
seem to be able to import it.)

Luckily, the protobuf library itself, which is all we really need, still
works, because if all else fails it can fall back to the python
implementation.  Well, for most people: for some, the fallback doesn't
work correctly!  (I suspect this has to do with an out of date
libpython, or perhaps a narrow build: the error was about an undefined
symbol `PyUnicodeUCS4_FromEncodedObject` but I haven't really
investigated further.)

Anyway, we can avoid the first problem by adding protobuf itself to
`sys.path` unconditionally (I made a separate directory with a symlink
to allow this) and the second problem by forcing fallback to python
protobufs if we are unable to import protobuf normally.  Just in case, I
also did a late-import of the whole datastore-translator, so users who
don't need it can avoid this.  (This is mostly for the benefit of
third-party users.)

Test plan:
- ran dev_appserver2.py, adding a log-line to the fallback code to
  ensure I don't see that (since I have a working protobuf
  implementation)
- forced grpc fallback to test that at least the workaround for the
  first problem works
- ask other folks who were seeing problems to test this and see that
  they are able to run `make serve` (which enables the translator
  automatically) without trouble.

Reviewers: csilvers

Subscribers: amos, colin, russell, yansong