Fix query cursor translation: a base64 story
This commit fixes a bug in the protobuf-over-HTTP version of the API,
where in certain cases it would crash when returning a query response.
The problem turns out to be a difference between how GAE and REST
serialize cursors, which, since both consider them opaque, wasn't a
problem, until proto came into the mix.

To wit: JSON-over-REST represents cursors as opaque strings; in practice
they appear to be base64 but really anything is fine as long as we're
consistent: all the client can do is pass them back to us (in the
context of the same query, even).  GAE internally uses
`datastore_query.Cursor`, but provides a standard urlsafe conversion, so
I had used that.  But it turns out that that urlsafe conversion uses
"urlsafe base64" (in the sense of python's `base64.urlsafe_b64encode`)
which is like ordinary base64 but with `+` and `/` replaced with `-` and
`_`.  Totally different!

As long as we used REST, this wasn't a problem, but when translating to
protobuf-over-REST, it turns out the base64 matters: the protobufs
represent these cursors as raw bytes, which only then get translated to
base64 for JSON.  So if our returned cursor in the JSON response isn't
valid base64, we (well really `google.protobuf.json_format`) can't
decode it to translate it to proto, and we crash.

Now, we use ordinary base64 everywhere.  While in the area, I refactored
a little so I didn't have to fix this in the 5 different places we
serialize and deserialize cursors; and added some documentation.

Test plan:
- `python/run_tests.py --frankenserver`
- follow Aasmund's reproduction steps, and ensure we no longer see the
  same bug, namely, pull this change into webapp, and:
  ```
  # in webapp root:
  make serve
  # in services/content_editing:
  rm -rf /tmp/frs-* # ensure we don't read from disk cache
  GCLOUD_PROJECT=khan-academy DATASTORE_EMULATOR_HOST=localhost:8001 ./gradlew run
  ```
  This should make it through english, and crash in spanish with some
  unrelated error, rather than getting a 500 from the translator.

Reviewers: csilvers

Subscribers: colin, aasmund, yansong