Implement non-transactional puts and deletes in datastore translator
In this commit, I implement the `Commit` request, which handles what GAE
calls puts and deletes, in `NON_TRANSACTIONAL` mode.  Most of the hard
work was already done in translating queries; this was just a matter of
wiring everything up with the right semantics.  I also finally
refactored out some of the input-format-checking that was getting
annoying to reimplement, which is the main bit of changes outside
`handlers.py`.

Test plan: `python/run_tests.py --frankenserver`

Reviewers: csilvers

Subscribers: colin, aasmund, yansong