Rewrite the datastore translator on top of Google's stub
Summary:
When originally planning the datastore emulator, we looked into using
Google's code for translating App Engine style calls to v1 REST calls.
Sadly, it was a part of the Cloud Datastore Emulator (a closed-source
Java binary) so we couldn't make use of it.  Little did we know they
also wrote code to translate the other direction -- the way I was
writing in the datastore translator!  I came across that code last week
while looking into how devappserver handles transactions.

So, now, this commit throws away most of our implementation in favor of
theirs.  While it's sad to throw away all my code, their implementation,
done with access to more information about the prod API's subtleties, is
surely better; and in any case it's already far more complete.  In
practice, we are likely to own any updates to it -- the code hasn't been
touched since 2015 -- but that's not so bad as this system is hopefully
temporary anyway, and the API is not changing very fast.  The upshot of
this is that this diff completes the implementation of the v1 REST API!
Absent any bugfixes, the datastore translator is now feature-complete.

Their implementation wasn't quite ready to use as-is.  First of all, it
wasn't wired up to anything other than an App Engine style apiproxy
stub, so much of the shims to handle HTTP and such remain.  (Their
implementation does use the protobuf form of the API, so the
proto-to-JSON translation is now inverted.)  Second, I had to fix
a number of bugs in their implementation, which clearly wasn't tested in
detail nor updated with the final version of the API.  The main sources
of difference are as follows:
- Project_id seems to have originally been required in a number of
  places where JSON doesn't include it at all.
- Some error codes have changed.
- Entity values can now be indexed -- this I just hacked since appengine
  can't handle it.
- The query operator HAS_PARENT was removed: there's only HAS_ANCESTOR
  now.
- The implementation was pretty careless in the difference between
  appengine style application IDs (`dev~khan-academy`) and GCP style
  project IDs (`khan-academy`).
All told, not a ton, which is why it's worth the change!  (And the diff,
although theoretically net-zero, is really quite net-negative: most of
the added lines are vendored code.)

This diff is pretty large, due to all the vendored code and all the
removals, so a description of the changes is in order:
- I ripped out most of the translation code (e.g. `translate_*.py`,
  `run_query.py`, and the API handler classes).
- Most of what was the `_DatastoreApiHandlerBase` is now the
  `DatastoreTranslatorHandler`; there are some changes (and some
  refactoring) but the basic idea -- do some format checking, delegate
  to [everything else] and handle errors.  (It may make sense to review
  this code as if it were new, even though some of it is not.)
- Since there wasn't a ton left, I inlined the remainder of `grpc.py`
  and `handlers.py` into `datastore_translator_server.py` and moved it
  to the parent directory (along with the tests).
- I vendored the `google-cloud-datastore` library into `lib`, and moved
  the genprotos to go with it -- more in the README of that directory.
- Some housekeeping due to the changes in where things live (`.arclint`,
  `python/run_tests.py`, and such).
- I fixed some things in the upstream code, as described above
  (this is most of the changes in `python/google/appengine/datastore`).

Test Plan:
To test these changes, I took three approaches:
- First, I ran the datastore client's "system tests" (what I would call
  "integration tests").  Sadly, this is a bit of a mess, but the
  docstring of `datastore_translator_server_test.py` describes how to do
  so, and the failures that remained.
- Second, I updated my own tests -- or at least the ones that tested the
  overall API rather than specific bits of my code.  (This makes for
  somewhat incomplete coverage, but it doesn't seem worth the effort to
  write many more tests at this time.)  I had to make some changes to
  support the differences between the two implementations:
  - We now skip asserting the returned entity-versions: their
    implementation is probably more like prod, but is nondeterministic.
  - The `meaning` and `indexUpdates` fields are now handled more
    accurately (or at least plausibly so, I didn't check).
  - The `moreResults` field on queries is arguably less accurate: in
    cases where this query has no more results but there might be more
    after a limit or end-cursor, we now return
    `MORE_RESULTS_AFTER_LIMIT` instead of trying to make a guess at
    which of the values is most correct.  This is sort of a bogus
    distinction in my opinion anyway.
  - A few bugs in my implementation and/or tests are now fixed.
- Finally, I tested this against our existing Kotlin services.
  - kotlin-routes was fairly easy: I just started it up and clicked
    around a bit.
  - content-editing required some work; I followed some, but not all, of
    the instructions in this doc:
    https://docs.google.com/document/d/1_n3K0W4BDoeOrWfjHi8cKJRx0cznSEdxLnirBEgEAFs/edit#
    Namely, I synced TAP data, gave myself permissions and passed the
    gandalf bridge, and then ran
    `GCLOUD_PROJECT=khan-academy DATASTORE_EMULATOR_HOST=localhost:8001 ./gradlew run`
    to start the service.  Finally, I went to the course editor (via the
    content editor as discussed in the doc), edited some content (it was
    a bit empty since I didn't bother syncing courses from prod) and
    refreshed the page to check that it was still there.
Other suggestions are welcome!

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: aasmund, colin

Differential Revision: https://phabricator.khanacademy.org/D55418