Fix a hang in checking whether grpc is importable
Summary:
We were passing `subprocess.PIPE` to `subprocess.call`, but not reading
from the pipe, which meant if the buffer filled up, the child process
would hang forever.  Now, instead, we pipe the output to devnull: we
don't intend to read it anyway, just to prevent the user from seeing it.
It's not clear exactly under what circumstances this hang would happen,
but it's not surprising that it would be a little arbitrary.

Note that this is a bug upstream, although it's right in and around a
bunch of code modified in frankenserver.

Test Plan: MDR, who was seeing the hang, tested this, and indeed the hang was fixed

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: mattdunnrankin

Differential Revision: https://phabricator.khanacademy.org/D57230