Fix string formatting in log messages
Summary:
While debugging something with Sign in with Apple I noticed that some log
messages show the raw format string and are not interpolating the
replacement parameters in.

This diff fixes two instances I noticed by adding parens around them.

Test Plan:
Run webapp and cause the assertion on either of the lines to be false.
Note that the message that is logged includes the name of the service who's
API proxy couldn't be found for.

Reviewers: benkraft, csilvers

Reviewed By: csilvers

Subscribers: #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D62315