Allow suppressing subprocess calls via an envvar.
Summary:
We do some work to have dev-appserver properly import grpc and
protocol-buffers.  This work takes about half a second, which is in
the noise for dev-appserver.

But it turns out that wrapper_util is imported (indirectly) by some of
our khan academy code, not just by dev-appserver.  In particular, all
our scripts call fix_sys_path() (to be able to import from
frankenserver) and that runs this code.  For our scripts, the 0.5
second overhead of these checks can be very noticeable.

But the checks aren't needed at all in that case, since we're not
running dev-appserver in that case (we're running a script!).  So this
envvar lets us communicate to tihs file whether we're being called
from fix_sys_path() or not.

It's possible we could do this in a more principled way than an
envvar, but the plumbing was difficult.  Ah well.

Issue: https://khanacademy.slack.com/archives/C3UDL9QN7/p1590301409346900

Test Plan: Fingers crossed

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D63487