Add logging for canonicalization
In https://github.com/Khan/frankenserver/pull/7 we patched the remote
API to canonicalize entities in order to allow our datastore migration
to continue.

However it appears that this mitigaiton didn't work. It's possible that
the mitigation didn't work because the code wasn't deployed in all the
right places. It's also possible that something is wrong with the mitigation
and it isn't doing what we think it should be doing.

This change should allow us to validate both these issues. The
presense of the logging messages that we're adding will show that the
new code is being run. Additionally we can compare the two log
messages - if they show that the entities are the same at these two
different points then the mitigation is working as designed.

Issue: None

Test plan:
- None
- Deploy to production and hope to see log messages appear