Add logging for canonicalization (#8)
## Summary:
In https://github.com/Khan/frankenserver/pull/7 we patched the remote
API to canonicalize entities in order to allow our datastore migration
to continue.

However it appears that this mitigation didn't work. It's possible that
the mitigation didn't work because the code wasn't deployed in all the
right places. It's also possible that something is wrong with the mitigation
and it isn't doing what we think it should be doing.

This change should allow us to validate the first possibility by adding log
messages showing that the new code is running.

We've decided for now to hold off on further logging due to not understanding
exactly what would be logged if we log the response. We want to avoid adding
PII or huge amounts of data to the logs.

We may do another PR with further logging if it turns out that our code IS being
run, but that the mitigation still isn't working.

Issue: None

## Test plan:
- None
- Deploy to production and hope to see log messages appear

Author: latteier

Reviewers: ipowell, dat-boris, eddavisson, mpolyak

Required Reviewers:

Approved By: ipowell, dat-boris, eddavisson

Checks:

Pull Request URL: https://github.com/Khan/frankenserver/pull/8