Ignore whether we can start the datastore emulator when running externally.
The dev-appserver makes sure it can start up the datastore emulator
when you run in "datastore emulator" mode.  However, if you pass in an
emulator host-port (via an envvar) then the dev-appserver never tries
to start up an emulator, so that check is spurious.  And in fact, it
was failing for us -- we don't have the proper binary installed in
frankenserver -- so it's good to bypass it.

Issue: https://khanacademy.atlassian.net/browse/INFRA-8969

Test plan:
I successfully ran
```
env DATASTORE_EMULATOR_HOST=localhost:8301 make serve-python HOST= FLAGS=--support_datastore_emulator=true
```
with this patch applied.